dtype fix

+++ basecaller/data.py
@@ -280,7 +280,7 @@ class EvalDataExtractor:
         self.files = [x for x in self.files if x.endswith('.fast5') or x.endswith('.signal')]
         if file_list is not None:
             self.files = [x for x in self.files if x.split('.')[0] in file_list]
-        self.output_types = (tf.float64, tf.int64, tf.string)
+        self.output_types = (tf.float32, tf.int64, tf.string)
         self.output_shapes = (tf.TensorShape([None, 300, 1]), tf.TensorShape([]), tf.TensorShape([]))
         self.current_file = 0
 
@@ -303,7 +303,7 @@ class EvalDataExtractor:
         else:
             raise ValueError(f"Format was {filename.split('.')[1]}, but it must be one of {', '.join(['.fast5', '.signal'])}.")
         signal, index = reader.read_for_eval(os.path.join(self.data_dir, filename))
-        signal = np.stack(signal)
+        signal = np.stack(signal).astype(np.float32)
         index = np.array(index)
         filenames = np.repeat(filename, index.shape[0])
         dataset = tf.data.Dataset.from_tensor_slices((
